declare _Noreturn usage() per C11; stop squelching warnings
There was a false positive warning on a construct like

  if (test) {...; d=...; ...} { usage(); }; foo(d);

because the d after the if statement is uninitialised if the test was
false. This isn't actually an issue because usage() calls exit(). By
declaring this, so the compiler knows, we squelch the false positive.

This allows removal of the -Wno-maybe-uninitialized compiler option.